refactor(llm): extract AnswerPlanner classification tables (211KB -> 139KB)
PR #427 §1.3 ("AnswerPlanner.ts is 211KB, no internal module split").

The PR proposes splitting out prompt TEMPLATES. Measured first, and that
premise is wrong: the 31 *_TEMPLATE constants are only ~16KB — 8% of the file.
Following the doc would have moved 8% and left AnswerPlanner at ~194KB.

Actual composition:
  planAnswer()           44KB (21%)
  31 *_PATTERNS tables   64KB (31%)
  everything else        the remainder

So the classification regex tables are the real extractable unit, and they are
pure data: arrays of regex literals, zero module-scope mutable state (`let`/
`var` at module scope: none), and no external consumer — all 24 importers
across the codebase pull exactly four symbols (planAnswer, isCodingAnswerType,
AnswerType, AnswerSource), none of them a pattern table.

Moved verbatim (comments included) to electron/llm/answerPlannerPatterns.ts and
imported back under identical names, so no reference inside planAnswer changes.
PEOPLE_OR_CONFLICT_OBJECT travels with them: it is interpolated INTO three of
the moved tables, and its own comment requires the three lists never drift
apart — typecheck caught this, which is why the helper moved rather than being
left behind and re-exported.

Kept as ONE patterns module rather than split per answer-type: a single
classifier cross-references the tables, so splitting further would add import
churn without improving navigability.

Verification: all 1017 regex literals present after the split (mechanically
diffed against the pre-split file — zero lost); 11 new tests including
behavioural cases routed through the REAL compiled planAnswer, so a dropped or
reordered pattern fails loudly; 192 existing planner tests (CodingContract,
ForbiddenLayers, AnswerStyle, ContextRoute, ...) pass; typecheck:electron and
build:electron clean. Confirmed end-to-end against LIVE DeepSeek v4 Flash
through the real pipeline — all 6 answer-shape invariants preserved vs the
pre-change baseline.

Not physically executed on Windows — shared platform-independent TypeScript,
no platform branches or packaging changes.
